refactor(core): 统一使用BigDecimal类型处理汇率- 修改Coin实体类中usdRate字段类型从double为BigDecimal - 更新AssetController中汇率设置逻辑，避免不必要的类型转换- 调整CheckExchangeRate任务中的汇率工厂设置方法 - 更新CoinExchangeFactoryConfig配置类中的汇率初始化逻辑 -优化CoinExchangeRate组件中的USDT/CNY汇率同步实现 - 引入validation-api依赖以增强数据校验能力- 升级unirest-java版本以支持新的HTTP客户端特性

+++ 01_bizzan_framework/01_bizzan_framework/core/src/main/java/com/bizzan/bitrade/entity/Coin.java
@@ -1,15 +1,23 @@
 package com.bizzan.bitrade.entity;
 
-import cn.afterturn.easypoi.excel.annotation.Excel;
-import lombok.Data;
-import lombok.ToString;
+import java.math.BigDecimal;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.persistence.Transient;
+
 import org.hibernate.validator.constraints.NotBlank;
 
 import com.bizzan.bitrade.constant.BooleanEnum;
 import com.bizzan.bitrade.constant.CommonStatus;
 
-import javax.persistence.*;
-import java.math.BigDecimal;
+import cn.afterturn.easypoi.excel.annotation.Excel;
+import lombok.Data;
+import lombok.ToString;
 
 /**
  * @author GS
@@ -64,7 +72,7 @@ public class Coin {
      */
     @Excel(name = "对美元汇率", orderNum = "1", width = 20)
     @Column(columnDefinition = "decimal(18,2) default 0.00 comment '美元汇率'")
-    private double usdRate;
+    private BigDecimal usdRate;
     /**
      * 是否支持rpc接口
      */
