+++ 01_bizzan_framework/01_bizzan_framework/ucenter-api/src/main/java/com/bizzan/bitrade/controller/AssetController.java
@@ -66,7 +66,7 @@ public class AssetController {
         wallets.forEach(wallet -> {
             CoinExchangeFactory.ExchangeRate rate = coinExchangeFactory.get(wallet.getCoin().getUnit());
             if (rate != null) {
-                wallet.getCoin().setUsdRate(rate.getUsdRate().doubleValue());
+                wallet.getCoin().setUsdRate(rate.getUsdRate());
                 wallet.getCoin().setCnyRate(rate.getCnyRate().doubleValue());
             } else {
                 log.info("unit = {} , rate = null ", wallet.getCoin().getUnit());
